Separating catalog op atoms from the rest of the stdlib
